the work of all methods is fixed, business logic is added

+++ src/main/java/com/Car_Rental_Spring/controller/requests/worker/WorkerUserUpdateRequest.java
@@ -9,6 +9,6 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class WorkerUserUpdateRequest extends WorkerUserCreateRequest {
-    private int workerId;
+    private Long workerId;
 }
